Move remaining lib/manageiq/release files into the bin scripts
